Shifted test_casts_from_tuple from test_kinematics.py to test_navigation.py.
